Split seed data into smaller modules and fix lin
